create PetsContext component to use React createContext and useContext hook to get pets data, instead of three listners in multiple screens, improved PorfileScreen's logic that only show My Pets and add button when myPets is not empty
