+++ PetDiaryApp/App.js
@@ -5,28 +5,31 @@ import { database } from "./firebase/firebaseSetup";
 
 import HomeScreen from "./screens/HomeScreen";
 import ProfileScreen from "./screens/ProfileScreen";
-import { colors } from "./styles";
 import EditLogScreen from "./screens/EditLogScreen";
+import { colors } from "./styles";
+import { PetsContextProvider } from "./components/PetsContext";
 
 const Stack = createNativeStackNavigator();
 
 export default function App() {
   //console.log(database);
   return (
-    <NavigationContainer>
-      <Stack.Navigator
-        screenOptions={{
-          headerStyle: { backgroundColor: colors.backgroundColor },
-        }}
-      >
-        <Stack.Screen
-          name="Home"
-          component={HomeScreen}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen name="Profile" component={ProfileScreen} />
-        <Stack.Screen name="Edit Log" component={EditLogScreen} />
-      </Stack.Navigator>
-    </NavigationContainer>
+    <PetsContextProvider>
+      <NavigationContainer>
+        <Stack.Navigator
+          screenOptions={{
+            headerStyle: { backgroundColor: colors.backgroundColor },
+          }}
+        >
+          <Stack.Screen
+            name="Home"
+            component={HomeScreen}
+            options={{ headerShown: false }}
+          />
+          <Stack.Screen name="Profile" component={ProfileScreen} />
+          <Stack.Screen name="Edit Log" component={EditLogScreen} />
+        </Stack.Navigator>
+      </NavigationContainer>
+    </PetsContextProvider>
   );
 }
