+++ PetDiaryApp/screens/AddLogScreen.js
@@ -12,13 +12,14 @@ import CustomTextInput from "../components/TextInput";
 import NameCheckbox from "../components/NameCheckbox";
 import { writeLogToDB } from "../firebase/firebasehelper";
 import { styles } from "../styles";
-import { collection, onSnapshot } from "firebase/firestore";
-import { database } from "../firebase/firebaseSetup";
 
 import { activitiesMenu } from "../constants";
+import { usePets } from "../components/PetsContext";
 import DropDownPicker from "react-native-dropdown-picker";
 
 const AddLogScreen = ({ navigation }) => {
+  const { myPets } = usePets();
+
   const returnForNoPets = () => {
     if (myPets.length === 0) {
       Alert.alert("You don't have any pet", "Please add a pet first.");
@@ -26,33 +27,12 @@ const AddLogScreen = ({ navigation }) => {
     }
   };
 
-  const [myPets, setMyPets] = useState([]);
-
   useEffect(() => {
-    // At Iteration 1, we are not using firebase authentication yet, so we are
-    // hardcoding the user to "testUser".
-    const q = collection(database, "PetDiary", "testUser", "pets");
-    const unsubscribe = onSnapshot(q, (querySnapshot) => {
-      if (!querySnapshot.empty) {
-        let pets = [];
-        querySnapshot.forEach((doc) => {
-          pets.push({ ...doc.data(), id: doc.id, isChecked: true });
-        });
-        setMyPets(pets);
-      } else {
-        setMyPets([]);
-        returnForNoPets();
-      }
+    const unsubscribe = navigation.addListener("focus", () => {
+      returnForNoPets();
     });
-    return () => unsubscribe();
-  }, []);
-
-  // useEffect(() => {
-  //   const unsubscribe = navigation.addListener("focus", () => {
-  //     returnForNoPets();
-  //   });
-  //   return unsubscribe;
-  // }, [navigation]);
+    return unsubscribe;
+  }, [navigation]);
 
   const [type, setType] = useState("");
   const [content, setContent] = useState("");
